refactor: move util function to external folder
